Multiple fixes throughout the code. Movement of multiple configs to new homes
