+++ SQF/dayz_code/Configs/UnleashedConfigs.hpp
@@ -1,0 +1,2 @@
+#include "CfgUnleashed\Homeopathic.hpp"
+#include "CfgUnleashed\Unleashed_weapons.hpp"
