+++ SQF/dayz_code/Configs/rscTitles.hpp
@@ -185,7 +185,7 @@ class RscDisplayMain : RscStandardDisplay
 		class DAYZ_Version : CA_Version
 		{
 			idc = -1;
-			text = "0.97-Devbuild-01252014";
+			text = "0.97-Devbuild-0126a2014";
 			y = "(SafeZoneH + SafeZoneY) - (1 - 0.95)";
 		};
 		delete CA_TitleMainMenu;
