+++ SQF/dayz_code/compile/agro_spawnZombies.sqf
@@ -52,4 +52,4 @@ _agent setVariable ["newDest",_player];
 */
 //Start behavior
 _id = [_position,_agent] execFSM "\z\AddOns\dayz_code\system\zombie_agent.fsm";
-diag_log ("ZASZ: " + str(_agent) + " Spawned At: " + str(_position));
+//diag_log ("ZASZ: " + str(_agent) + " Spawned At: " + str(_position));
