+++ SQF/dayz_code/init/compiles.sqf
@@ -268,6 +268,84 @@ if (!isDedicated) then {
 		private ["_dikCode", "_handled","_displayg"];
 		_dikCode = _this select 1;
 		_handled = false;
+		if (_dikCode in[0x02,0x03,0x04,0x58,0x57,0x44,0x43,0x42,0x41,0x40,0x3F,0x3E,0x3D,0x3C,0x3B,0x0B,0x0A,0x09,0x08,0x07,0x06,0x05]) then {
+			_handled = true;
+		};
+
+		// esc
+		if (_dikCode == 0x01) then {
+			DZE_cancelBuilding = true;
+		};
+		
+		// surrender 
+		if (_dikCode in actionKeys "Surrender") then {
+			
+			_vehicle = vehicle player;
+			_inVehicle = (_vehicle != player);
+			_onLadder =	(getNumber (configFile >> "CfgMovesMaleSdr" >> "States" >> (animationState player) >> "onLadder")) == 1;
+			_canDo = (!r_drag_sqf and !r_player_unconscious and !_onLadder and !_inVehicle);
+			
+			if (_canDo and !DZE_Surrender and !(player isKindOf  "PZombie_VB")) then {
+				DZE_Surrender = true;
+				_dropPrimary = false;
+				_dropSecondary = false;
+
+				_primaryWeapon = primaryWeapon player;
+				if (_primaryWeapon != "") then {_dropPrimary = true;};
+				_secondaryWeapon = "";
+				{
+					if ((getNumber (configFile >> "CfgWeapons" >> _x >> "Type")) == 2) exitWith {
+							_secondaryWeapon = _x;
+					};
+				} forEach (weapons player);
+				if (_secondaryWeapon != "") then {_dropSecondary = true;};
+
+				if (_dropPrimary or _dropSecondary) then {
+					player playActionNow "PutDown";
+					_iPos = getPosATL player;
+					_radius = 1;
+					_item = createVehicle ["WeaponHolder", _iPos, [], _radius, "CAN_COLLIDE"];
+					_item setposATL _iPos;
+					if (_dropPrimary) then {
+						_iItem = _primaryWeapon;
+						_removed = ([player,_iItem,1] call BIS_fnc_invRemove);
+						if (_removed == 1) then {
+							_item addWeaponCargoGlobal [_iItem,1];
+						};
+					};
+					if (_dropSecondary) then {
+						_iItem = _secondaryWeapon;
+						_removed = ([player,_iItem,1] call BIS_fnc_invRemove);
+						if (_removed == 1) then {
+							_item addWeaponCargoGlobal [_iItem,1];
+						};
+					};
+					player reveal _item;
+				};
+
+				// set publicvariable that allows other player to access gear
+				player setVariable ["DZE_Surrendered", true, true];
+				// surrender animation
+				player playMove "AmovPercMstpSsurWnonDnon";
+			};
+			_handled = true;
+		};
+		
+		
+		if (_dikCode in actionKeys "MoveForward") exitWith {r_interrupt = true; if (DZE_Surrender) then {call dze_surrender_off};};
+		if (_dikCode in actionKeys "MoveLeft") exitWith {r_interrupt = true; if (DZE_Surrender) then {call dze_surrender_off};};
+		if (_dikCode in actionKeys "MoveRight") exitWith {r_interrupt = true; if (DZE_Surrender) then {call dze_surrender_off};};
+		if (_dikCode in actionKeys "MoveBack") exitWith {r_interrupt = true; if (DZE_Surrender) then {call dze_surrender_off};};
+
+				//Prevent exploit of drag body
+		if ((_dikCode in actionKeys "Prone") and r_drag_sqf) then { force_dropBody = true; };
+		if ((_dikCode in actionKeys "Crouch") and r_drag_sqf) then { force_dropBody = true; };
+		
+		_shift = 	_this select 2;
+		_ctrl = 	_this select 3;
+		_alt =		_this select 4;
+		
+		
 		if (_dikCode in (actionKeys "GetOver")) then {
 			// this prevents the player zombie from performing the getover action as it will cause pz to just get stuck standing. 
 			// but will cause side effect of auto run/walk, perform attack to stop.
@@ -291,9 +369,7 @@ if (!isDedicated) then {
 			_nill = execvm "\z\addons\dayz_code\actions\playerstats.sqf";
 		};
 		*/
-		//Prevent exploit of drag body
-		if ((_dikCode in actionKeys "Prone") and r_drag_sqf) then { force_dropBody = true; };
-		if ((_dikCode in actionKeys "Crouch") and r_drag_sqf) then { force_dropBody = true; };
+
 
 		if (_dikCode in actionKeys "MoveLeft") then {r_interrupt = true};
 		if (_dikCode in actionKeys "MoveRight") then {r_interrupt = true};
@@ -321,6 +397,7 @@ if (!isDedicated) then {
 			_nill = execvm "\z\addons\dayz_code\actions\playerstats.sqf";
 		};
 		
+
 		// numpad 8 0x48 now pgup 0xC9 1
 		if ((_dikCode == 0xC9 and (!_alt or !_ctrl)) or (_dikCode in actionKeys "User15")) then {
 			DZE_Q = true;
@@ -350,9 +427,6 @@ if (!isDedicated) then {
 			DZE_Z_ctrl = true;
 		};
 
-
-
-
 		// numpad 4 0x4B now Q 0x10
 		if (_dikCode == 0x10 or (_dikCode in actionKeys "User17")) then {
 			DZE_4 = true;
@@ -365,12 +439,6 @@ if (!isDedicated) then {
 		if (_dikCode == 0x39 or (_dikCode in actionKeys "User19")) then {
 			DZE_5 = true;
 		};
-
-		// esc
-		if (_dikCode == 0x01) then {
-			DZE_cancelBuilding = true;
-		};
-		
 		
 		if (_dikCode == 0x15 ) then {
 			if (!dayz_isSwimming and !dialog) then {
@@ -380,11 +448,6 @@ if (!isDedicated) then {
 			_values = ["val1", "val2", "val3", "etc"];
 			[_keys, _values] call updateUI;
 			_handled = true;
-		};
-		
-		if ((_dikCode == 0x3E or _dikCode == 0x0F or _dikCode == 0xD3) and (time - dayz_lastCheckBit > 10)) then {
-			dayz_lastCheckBit = time;
-			call dayz_forceSave;
 		};
 		
 		if (_dikCode == 0xB8 or _dikCode == 0x38 or _dikCode == 0x3E or _dikCode == 0x2A or _dikCode == 0x36 or _dikCode == 0x01) then {
@@ -521,7 +584,11 @@ if (!isDedicated) then {
 			_control ctrlCommit 0;
 		};
 	};
-
+	
+	dze_surrender_off = {
+		player setVariable ["DZE_Surrendered", false, true];
+		DZE_Surrender = false;	
+	};
 
 	gear_ui_init = {
 		private["_control","_parent","_menu","_dspl","_grpPos"];
