+++ SQF/dayz_server/DZAI/compile/heli_airlanding.sqf
@@ -36,8 +36,7 @@ _weapongrade = DZAI_heliEquipType call DZAI_getWeapongrade;
 		0 = [_x, _weapongrade] call DZAI_setSkills;
 		0 = [_x, _weapongrade] spawn DZAI_autoRearm_unit;
 		_x setVariable ["unconscious",false];
-		_x setVariable ["unithealth",[12000,0,0,false,false]];
-		_x setHit["hands",0];
+		_x setVariable ["unithealth",[12000,0,false]];
 		_x setHit["legs",0];
 		unassignVehicle _x;
 	};
