+++ SQF/dayz_server/DZAI/init/world_classname_configs/default/default_classnames.sqf
@@ -23,7 +23,6 @@ if !(DZAI_dynamicWeaponList) then {
 	DZAI_Rifles2 = ["M16A2","M16A2GL","M249_DZ","AK_74","M4A1_Aim","AKS_74_kobra","AKS_74_U","AK_47_M","M24","SVD_CAMO","M1014","DMR_DZ","M4A1","M14_EP1","Remington870_lamp","M240_DZ","M4A1_AIM_SD_camo","M16A4_ACG","M4A1_HWS_GL_camo","Mk_48_DZ","M4A3_CCO_EP1","Sa58V_RCO_EP1","Sa58V_CCO_EP1","M40A3","Sa58P_EP1","Sa58V_EP1"]; //Weapongrade 2 rifles
 	DZAI_Rifles3 = ["FN_FAL","FN_FAL_ANPVS4","Mk_48_DZ","M249_DZ","BAF_L85A2_RIS_Holo","G36C","G36C_camo","G36A_camo","G36K_camo","AK_47_M","AKS_74_U","M14_EP1","bizon_silenced","DMR_DZ","RPK_74"]; //Weapongrade 3 rifles
 	
-	
 	DZAI_weaponsInitialized = true; //Do not edit this.
 };
 //Note: Custom rifle tables can be defined below this line (DZAI_Rifles4 -DZAI_Rifles9). Custom rifle tables can only be used with custom-defined spawns (spawns created using the DZAI_spawn function).
