+++ SQF/dayz_server/DZAI/spawn_functions/spawn_vehPatrol.sqf
@@ -48,9 +48,9 @@ for "_i" from 1 to _amount do {
 	_driver moveInDriver _vehicle;
 	0 = [_driver,_weapongrade] call DZAI_setSkills;
 	0 = [_driver,_weapongrade] call DZAI_setupLoadout;
-	_driver setVariable ["unithealth",[12000,0,0,false,false]];
+	_driver setVariable ["unithealth",[12000,0,false]];
 	_driver setVariable ["unconscious",true];	//Prevent AI heli crew from being knocked unconscious
-	_driver setVariable ["DZAI_deathTime",time];
+	//_driver setVariable ["DZAI_deathTime",time];
 	if (!(_driver hasWeapon "NVGoggles")) then {
 		_driver addWeapon "NVGoggles";
 		_driver setVariable ["removeNVG",1];
@@ -64,9 +64,9 @@ for "_i" from 1 to _amount do {
 		[_cargo] joinSilent _unitGroup;
 		0 = [_cargo,_weapongrade] call DZAI_setSkills;
 		0 = [_cargo,_weapongrade] call DZAI_setupLoadout;
-		_cargo setVariable ["unithealth",[12000,0,0,false,false]];
+		_cargo setVariable ["unithealth",[12000,0,false]];
 		_cargo setVariable ["unconscious",true];	//Prevent AI heli crew from being knocked unconscious
-		_cargo setVariable ["DZAI_deathTime",time];
+		//_cargo setVariable ["DZAI_deathTime",time];
 		if (!(_cargo hasWeapon "NVGoggles")) then {
 			_cargo addWeapon "NVGoggles";
 			_cargo setVariable ["removeNVG",1];
@@ -84,9 +84,9 @@ for "_i" from 1 to _amount do {
 		[_gunner] joinSilent _unitGroup;
 		0 = [_gunner,_weapongrade] call DZAI_setSkills;
 		0 = [_gunner,_weapongrade] call DZAI_setupLoadout;
-		_gunner setVariable ["unithealth",[12000,0,0,false,false]];
+		_gunner setVariable ["unithealth",[12000,0,false]];
 		_gunner setVariable ["unconscious",true];	//Prevent AI heli crew from being knocked unconscious
-		_gunner setVariable ["DZAI_deathTime",time];
+		//_gunner setVariable ["DZAI_deathTime",time];
 		if (!(_gunner hasWeapon "NVGoggles")) then {
 			_gunner addWeapon "NVGoggles";
 			_gunner setVariable ["removeNVG",1];
